test(cli): address integration command review
